server/mediapipe-backend/requirements.txt
absl-py==1.0.0
attrs==21.4.0
cycler==0.11.0
fonttools==4.31.1
grpcio==1.44.0
grpcio-tools==1.44.0
kiwisolver==1.4.0
matplotlib==3.5.1
mediapipe==0.8.9.1
numpy==1.22.3
opencv-contrib-python==4.5.5.64
packaging==21.3
Pillow==9.0.1
protobuf==3.19.4
pyparsing==3.0.7
python-dateutil==2.8.2
six==1.16.0
